moved useErrorFocus tests
